Workaround for PacketBuffer bug.
There exist a bug in the video_coding::PacketBuffer which triggers when a
frame is the same size as the buffer. A trivial workaround is to increase
the start size to something big so that this never happens in practice.

The bug has been fixed but we still want to test the workaround in ToT,
which is why this CL exist.

BUG=webrtc:8028, chromium:752886
R=stefan@webrtc.org

Review-Url: https://codereview.webrtc.org/2994093002 .
Cr-Original-Commit-Position: refs/heads/master@{#19308}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 3bf97cf0602f150e05b5424735cced520aa0cfef